Recover weekly from garbled all-models line; move test to own file
Address review: apply the tolerant all-models match in the weekly percent/reset extraction too, so a capture where only the corrupted "all modls" label survived still populates the Weekly limit instead of dropping it. Move the new cases into their own test type to stay under type_body_length.
